requirements.txt
Django==2.1.4
pytz==2018.7
defusedxml==0.5.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.1.4
django-import-export==1.1.0
et-xmlfile==1.0.1
gunicorn==19.9.0
jdcal==1.4
numpy==1.15.4
odfpy==1.4.0
openpyxl==2.5.12
pandas==0.23.4
psycopg2==2.7.6.1
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
six==1.12.0
tablib==0.12.1
unicodecsv==0.14.1
whitenoise==4.1.2
xlrd==1.2.0
xlwt==1.3.0
